+++ packages/design-system-core/src/lib/createSystemProp/createSystemProp.ts
@@ -1,7 +1,7 @@
 import flatten from 'lodash/flatten';
 import get from 'lodash/get';
-import { useCurrentTheme } from '@class101/design-system-theme';
 import { isDefined } from '@class101/design-system-utils';
+import { useCurrentTheme } from '../ThemeProvider';
 import { useBuildStyle } from '../useBuildStyle';
 import type { SystemProp, SystemPropConfig } from './type';
